changed ResetFilingPathBehavior to BasePathBehavior
